+++ plugins/svg/src/SvgFeatureRenderer/components/FeatureGlyph.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react'
 
-import type { DisplayModel } from './util'
+import type { DisplayModel } from './types'
 import type { AnyConfigurationModel } from '@jbrowse/core/configuration'
 import type { Feature, Region } from '@jbrowse/core/util'
 import type { SceneGraph } from '@jbrowse/core/util/layouts'
@@ -31,10 +31,10 @@ const FeatureGlyph = observer(function (props: {
   const featureLayout = rootLayout.getSubRecord(String(feature.id()))
   if (!featureLayout) {
     return null
+  } else {
+    const { GlyphComponent } = featureLayout.data || {}
+    return <GlyphComponent featureLayout={featureLayout} {...props} />
   }
-  const { GlyphComponent } = featureLayout.data || {}
-
-  return <GlyphComponent featureLayout={featureLayout} {...props} />
 })
 
 export default FeatureGlyph
